requirements.txt
Flask
Flask-Cors
python-dotenv
pymysql
flask-mysqldb
pyjwt
typing
dataclasses